all centavos replaced with peso

+++ source/Ssi.TrackTruck/Ssi.TrackTruck.Bussiness/Employees/EmployeeSalary.cs
@@ -5,8 +5,8 @@ namespace Ssi.TrackTruck.Bussiness.Employees
     public class EmployeeSalary
     {
         public DbEmployee Employee { get; set; }
-        public long TotalAllowance { get; set; }
-        public long TotalSalary { get; set; }
-        public long TotalPayable { get { return TotalAllowance + TotalSalary; } }
+        public double TotalAllowance { get; set; }
+        public double TotalSalary { get; set; }
+        public double TotalPayable { get { return TotalAllowance + TotalSalary; } }
     }
 }
